fix, test: order F params correctly, add runtime check for valid OM
Fix the order of the F params, following order in Report.sso PARAMETER table. Adds  runtime check that 1) model can run and 2)init_F and F_rate params are the assumed length and in the correct order. passes tests in test-initOM.R. Could potentially expand this check to other manual param changes in the ss.par file to check assumptions.