fix: adding new values from OM run to EM data set
Discovered that sd's/Sample sizes are sometimes different; this is because the ss bootstrapping module adjust the sd/ sample size to include extra SD and the input variance from ctl file. Still need to think more about this wrt the EM - probably should turn off extra SD and input variance in the EM based on this?