fix: bug fixes for interim catch forecast
This extends the EM forecast period to make sure enough catches are projected for the reference files. Also fixes a few other bugs and adds a maximum catch adjustment range of 0.25 to 2 times reference. This will likely not be needed but helps reign in the issues while projected indicies are still acting up.