fix, dev: return devs, abs, base params
Realized the convert_future_om_list_to_df function was only returning abs values. It should be returning devs from the base parameter value. However, I also realized it may be useful to return all 3 values in a list of dataframes so they can be examined and used as needed. Also added to the function description and some empty tests that should be implemented.